# backend/app/api/auth.py
from datetime import timedelta, datetime
from typing import Any
from fastapi import APIRouter, Depends, HTTPException, status, Request
from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.future import select
from jose import jwt, JWTError
from passlib.context import CryptContext

from app.core.database import get_db
from app.core.config import settings
from app.models.user import User
from app.schemas.user import UserCreate, UserResponse, Token, UserUpdate
from app.core.limiter import limiter

router = APIRouter()
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"{settings.API_V1_STR}/auth/login")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(subject: str, expires_delta: timedelta = None) -> str:
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm="HS256")
    return encoded_jwt

from sqlalchemy.orm import selectinload

async def get_current_user(db: AsyncSession = Depends(get_db), token: str = Depends(oauth2_scheme)) -> User:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    
    result = await db.execute(
        select(User)
        .where(User.id == user_id)
        .options(selectinload(User.business_profile))
    )
    user = result.scalars().first()
    if user is None:
        raise credentials_exception
    return user

def require_feature(feature_key: str):
    async def dependency(current_user: User = Depends(get_current_user)) -> User:
        if not current_user.business_profile:
            raise HTTPException(status_code=404, detail="Business profile not found")
        
        # Local import to prevent circular imports
        from app.api.business import DEFAULT_FEATURES_CONFIG
        cfg = current_user.business_profile.features_config or DEFAULT_FEATURES_CONFIG
        
        if not cfg.get(feature_key, {}).get("enabled", False):
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN, 
                detail=f"El módulo '{feature_key}' no está habilitado para esta cuenta."
            )
        return current_user
    return dependency

def require_any_feature(feature_keys: list):
    async def dependency(current_user: User = Depends(get_current_user)) -> User:
        if not current_user.business_profile:
            raise HTTPException(status_code=404, detail="Business profile not found")
        
        # Local import to prevent circular imports
        from app.api.business import DEFAULT_FEATURES_CONFIG
        cfg = current_user.business_profile.features_config or DEFAULT_FEATURES_CONFIG
        
        has_any = any(cfg.get(key, {}).get("enabled", False) for key in feature_keys)
        if not has_any:
            features_label = " o ".join(f"'{k}'" for k in feature_keys)
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN, 
                detail=f"Ninguno de los módulos requeridos ({features_label}) está habilitado para esta cuenta."
            )
        return current_user
    return dependency

@router.get("/me", response_model=UserResponse)
async def get_user_me(current_user: User = Depends(get_current_user)) -> Any:
    return current_user

@router.patch("/me", response_model=UserResponse)
async def update_user_me(
    user_in: UserUpdate,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
) -> Any:
    if user_in.email:
        result = await db.execute(select(User).where(User.email == user_in.email))
        user = result.scalars().first()
        if user and user.id != current_user.id:
            raise HTTPException(status_code=400, detail="Email already registered")
        current_user.email = user_in.email
    
    if user_in.password:
        current_user.hashed_password = get_password_hash(user_in.password)
    
    db.add(current_user)
    await db.commit()
    # Refresh with eager load
    result = await db.execute(
        select(User)
        .where(User.id == current_user.id)
        .options(selectinload(User.business_profile))
    )
    return result.scalars().first()

@router.post("/register", response_model=UserResponse)
async def register(user_in: UserCreate, db: AsyncSession = Depends(get_db)) -> Any:
    result = await db.execute(select(User).where(User.email == user_in.email))
    user = result.scalars().first()
    if user:
        raise HTTPException(
            status_code=400,
            detail="The user with this email already exists in the system.",
        )
    user = User(
        email=user_in.email,
        hashed_password=get_password_hash(user_in.password),
        role="member",
        is_admin=False,
        is_active=True
    )
    db.add(user)
    await db.commit()
    # Refresh with eager load
    result = await db.execute(
        select(User)
        .where(User.id == user.id)
        .options(selectinload(User.business_profile))
    )
    return result.scalars().first()

@router.post("/login", response_model=Token)
@limiter.limit("5/minute")
async def login(
    request: Request,
    db: AsyncSession = Depends(get_db), 
    form_data: OAuth2PasswordRequestForm = Depends()
) -> Any:
    result = await db.execute(select(User).where(User.email == form_data.username))
    user = result.scalars().first()
    if not user or not verify_password(form_data.password, user.hashed_password):
        raise HTTPException(status_code=400, detail="Incorrect email or password")
    
    access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    return {
        "access_token": create_access_token(user.id, expires_delta=access_token_expires),
        "token_type": "bearer",
    }